+++ release/fv/fv_dakelh/HISTORY.md
@@ -1,8 +1,14 @@
 Dakelh Change History
 ============================
+9.1.6 (4 Jul 2024)
+-------------------
+* Simplified missed e with acute in mobile layout
+* Added alternate way to type / in Google Workspace
+* Tidied documentation
+
 9.1.5 (12 Feb 2024)
 -------------------
-* Cimplified to single code points were possible
+* Simplified to single code points were possible
 * Cleaned up rules
 * Centered mobile keyboard